requirements.txt
streamlit
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
transformers
torch